backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.5.2
ultralytics
opencv-python-headless
numpy
httpx==0.27.0
# torch is installed by ultralytics
